Moved CommonParametersTrait into a parameters namespace
