+++ src/Smrtr/MysqlVersionControl/Command/Teardown.php
@@ -2,11 +2,12 @@
 
 namespace Smrtr\MysqlVersionControl\Command;
 
+use Smrtr\MysqlVersionControl\Command\Parameters\CommonParametersTrait;
+use Smrtr\MysqlVersionControl\Receiver\Teardown as TeardownReceiver;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Smrtr\MysqlVersionControl\Receiver\Teardown as TeardownReceiver;
 
 /**
  * Class Teardown is a symfony console command that removes all tables from the database for the given environment.
